pyproject.toml
[project]
name = "effective-giggle"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.24.0",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "tiktoken>=0.6",
    "requests>=2.31",
    "openai-agents>=0.1.0",
    "pydantic-settings>=2.2"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["effective_giggle*"]
exclude = ["vendor_docs*"]
